# backend/swagger/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/auth/oidc/callback:
    get:
      consumes:
      - application/json
      description: Handles OIDC login callback.
      parameters:
      - description: OIDC code
        in: query
        name: code
        type: string
      - description: OIDC state
        in: query
        name: state
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OIDC callback result
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: OIDC callback
      tags:
      - Auth
...
swagger: "2.0"